speakeasy_emulator/requirements.txt
fastapi
uvicorn
python-multipart
speakeasy
yara-python
setuptools<82